+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.4] - 2022-01-01
+
 ### Added
 
 - Added `MessageDescriptor::map_entry_key_field` and `MessageDescriptor::map_entry_value_field` to help with inspecting message types.
@@ -65,7 +67,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release, including support for inspecting message types at runtime.
 
-[unreleased]: https://github.com/andrewhickman/prost-reflect/compare/0.3.3...HEAD
+[unreleased]: https://github.com/andrewhickman/prost-reflect/compare/0.3.4...HEAD
+[0.3.4]: https://github.com/andrewhickman/prost-reflect/compare/0.3.3...0.3.4
 [0.3.3]: https://github.com/andrewhickman/prost-reflect/compare/0.3.2...0.3.3
 [0.3.2]: https://github.com/andrewhickman/prost-reflect/compare/0.3.1...0.3.2
 [0.3.1]: https://github.com/andrewhickman/prost-reflect/compare/0.3.0...0.3.1